Should be all good ✨

+++ philo/src/table/clean_table.c
@@ -6,11 +6,12 @@
 /*   By: hanmpark <hanmpark@student.42nice.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/05/10 15:00:29 by hanmpark          #+#    #+#             */
-/*   Updated: 2023/05/10 15:54:01 by hanmpark         ###   ########.fr       */
+/*   Updated: 2023/05/11 15:08:36 by hanmpark         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "simulation.h"
+#include "errors.h"
 
 static void	destroy_mutexes(t_table *table, int nbr_lock, int nbr_fork)
 {
@@ -22,15 +23,16 @@ static void	destroy_mutexes(t_table *table, int nbr_lock, int nbr_fork)
 	i = -1;
 	while (++i < nbr_fork)
 		pthread_mutex_destroy(&table->fork[i]);
-	pthread_mutex_destroy(&table->end_sim_lock);
-	pthread_mutex_destroy(&table->print);
+	pthread_mutex_destroy(&table->sim_lock);
+	pthread_mutex_destroy(&table->print_lock);
 }
 
 void	clean_table(t_table *table, int nbr_lock, int nbr_fork)
 {
 	if (!table)
 		return ;
-	destroy_mutexes(table, nbr_lock, nbr_fork);
+	if (nbr_lock != 0 || nbr_fork != 0)
+		destroy_mutexes(table, nbr_lock, nbr_fork);
 	if (table->fork != NULL)
 		free(table->fork);
 	if (table->philo != NULL)
